requirements.txt
allure-pytest==2.8.6
selenium==2.53.6